[bookmark_1_0_0]  Fix when selecting bookmarks
[memo_1_0_0]
 Disabled edit field while loading content

[system_1_0_0]
 Minor fixes.


git-svn-id: http://dev.deskportal.net/source/live@70 ffa8ddc6-ed64-449d-89c4-29ad718b3e2b

+++ bookmark/1/0/0/server/php/module/group.php
@@ -29,9 +29,7 @@
 			if(!$user->valid) return false;
 
 			$database = $system->database('user', __METHOD__, $user);
-
-			if(!$database->success) return false;
-			if(!$database->begin()) return false; #Keep the category and relation together
+			if(!$database->success || !$database->begin()) return false; #Keep the category and relation together
 
 			$query = $database->prepare("DELETE FROM {$database->prefix}category WHERE id = :id AND user = :user");
 			$query->run(array(':id' => $id, ':user' => $user->id));
